# docs/swagger-old.yaml
openapi: 3.0.0
info:
  title: APIServiceGo
  version: 1.0.0
  description: EV Charging API for searching charging connectors, getting estimates, managing orders, and processing payments for electric vehicle charging sessions.
servers:
  - url: http://{{localhost}}
paths:
...
  /v1/orders/{order_id}/support:
    get:
      operationId: getOrderSupport
      summary: Get Order Support
      tags:
        - Support
      parameters:
        - name: X-Transaction-Id
          in: header
          required: true
          schema:
            type: string
        - name: X-Bpp-Id
          in: header
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupportResponse"
              example:
                order:
                  id: order-bpp-789012
                  status: CANCELLED
                  mode: RESERVATION
                name: BlueCharge Support Team
                phone: "18001080"
                email: support@bluechargenet-aggregator.io
                url: https://support.bluechargenet-aggregator.io/ticket/SUP-20250730-001
                hours: Mon–Sun 24/7 IST
                channels:
                  - phone
                  - email
                  - web
                  - chat
        "401":
          description: Unauthorized - Invalid or missing authentication token.
          headers:
            X-Transaction-Id:
              required: true
              schema:
                type: string
                description: Transaction identifier for request tracking
            X-Bpp-Id:
              required: true
              schema:
                type: string
                description: BPP identifier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error:
                  code: "UNAUTHORIZED"
                  message: "Invalid or missing authentication token."
        "404":
          description: Not Found - Order not found.
          headers:
            X-Transaction-Id:
              required: true
              schema:
                type: string
                description: Transaction identifier for request tracking
            X-Bpp-Id:
              required: true
              schema:
                type: string
                description: BPP identifier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error:
                  code: "NOT_FOUND"
                  message: "Order not found."
        "500":
          description: Internal Server Error - Server error occurred while processing the request.
          headers:
            X-Transaction-Id:
              required: true
              schema:
                type: string
                description: Transaction identifier for request tracking
            X-Bpp-Id:
              required: true
              schema:
                type: string
                description: BPP identifier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error:
                  code: "INTERNAL_ERROR"
                  message: "Server error occurred while processing the request."
    parameters:
      - name: order_id
        in: path
        required: true
        schema:
          type: string
components:
  schemas:
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              description: Error code
              example: "INVALID_REQUEST"
            message:
              type: string
              description: Human-readable error message
              example: "Invalid request parameters"
            details:
              type: object
              description: Additional error details
              additionalProperties: true
...
    OrderInfo:
      type: object
      description: Order information
      properties:
        id:
          type: string
          description: Order identifier
        mode:
          type: string
          description: Order mode (e.g., reservation, RESERVATION)
        status:
          type: string
          description: Order status (e.g., quoted_price, ACTIVE, CANCELLED, COMPLETED)
...
    SupportResponse:
      type: object
      description: Support response
      properties:
        order:
          $ref: "#/components/schemas/OrderInfo"
        name:
          type: string
          description: Support team name
        phone:
          type: string
          description: Support phone number
        email:
          type: string
          format: email
          description: Support email address
        url:
          type: string
          format: uri
          description: URL to support portal or ticket
        hours:
          type: string
          description: Support hours of operation
        channels:
          type: array
          description: Available support channels
          items:
            type: string
            description: Support channel (e.g., phone, email, web, chat)
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
security:
  - BearerAuth: []
...